[DPE-8884] Implement charm refreshing
