+++ tests/spread/test_refresh.py/task.yaml
@@ -1,0 +1,5 @@
+summary: test_refresh.py
+execute: |
+  tox run -e integration-refresh -- --alluredir="$SPREAD_TASK/allure-results" --model-name=test-cas --keep-models
+artifacts:
+  - allure-results
