+++ tox.ini
@@ -67,7 +67,7 @@ commands =
                  {[vars]tests_path}/unit
     poetry run coverage report
 
-[testenv:integration-{authentication,charm,config,restart,multinode,scaling,tls,cos,provider,multicluster,network}]
+[testenv:integration-{authentication,charm,config,restart,multinode,scaling,tls,cos,provider,multicluster,network,refresh}]
 description = Run integration tests
 pass_env =
     CI
@@ -83,6 +83,7 @@ set_env =
     provider: TESTFILE=test_provider.py
     multicluster: TESTFILE=test_multicluster.py
     network: TESTFILE=test_network.py
+    refresh: TESTFILE=test_refresh.py    
 commands_pre =
     poetry install --only integration
 commands =
